+++ src/features/basic/index.ts
@@ -49,6 +49,7 @@ import './lm-item-icons/lm-item-icons';
 import './mat-linkify-category';
 import './mtra-auto-focus-amount';
 import './mtra-transfer-on-enter';
+import './mu-fix-sector-names';
 import './nots-material-ticker';
 import './nots-notification-type-label';
 import './nots-ship-arrival-inventory';